pyproject.toml
[tool.poetry]
name = "citas-medicas-backend"
version = "0.1.0"
description = "Backend Django para plataforma de citas médicas"
authors = ["Equipo"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12,<=3.14"
django = "^5.0"
djangorestframework = "^3.15"
django-environ = "^0.11"
psycopg2-binary = "^2.9"
django-cors-headers = "^4.4"
celery = "^5.4"
redis = "^5.0"
twilio = "^9.0"
djangorestframework-stubs = "^3.16.6"
djangorestframework-simplejwt = "^5.3"
drf-yasg = "^1.21"
boto3 = "^1.35"
azure-communication-email = "^1.1"
azure-core = "^1.30"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.8"
black = "^24.10"
ruff = "^0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

